+++ packages/qwik/src/index.ts
@@ -3,5 +3,6 @@ export * from './adapters';
 export * from './components';
 export * from './hooks';
 export * from './methods';
+export * from './transformation';
 export * from './types';
 export * from './validation';