+++ packages/solid/src/index.ts
@@ -2,5 +2,6 @@ export * from './adapters';
 export * from './components';
 export * from './methods';
 export { createForm, createFormStore } from './primitives';
+export * from './transformation';
 export * from './types';
 export * from './validation';