+++ packages/solid/src/utils/handleFieldEvent.ts
@@ -1,24 +1,28 @@
 import { batch } from 'solid-js';
 import type {
+  FieldEvent,
   FieldPath,
   FieldPathValue,
   FieldValues,
   FormStore,
   InternalFieldStore,
   ResponseData,
+  ValidationMode,
 } from '../types';
 import { updateFieldDirty } from './updateFieldDirty';
 import { validateIfRequired } from './validateIfRequired';
 
 /**
- * Updates the value state of a field.
+ * Handles the input, change and blur event of a field.
  *
  * @param form The form of the field.
  * @param field The store of the field.
  * @param name The name of the field.
- * @param value The new value of the field.
+ * @param event The event of the field.
+ * @param validationModes The modes of the validation.
+ * @param inputValue The value of the input.
  */
-export function updateFieldValue<
+export function handleFieldEvent<
   TFieldValues extends FieldValues,
   TResponseData extends ResponseData,
   TFieldName extends FieldPath<TFieldValues>
@@ -26,11 +30,18 @@ export function updateFieldValue<
   form: FormStore<TFieldValues, TResponseData>,
   field: InternalFieldStore<TFieldValues, TFieldName>,
   name: TFieldName,
-  value: FieldPathValue<TFieldValues, TFieldName>
-): void {
+  event: FieldEvent,
+  validationModes: Exclude<ValidationMode, 'submit'>[],
+  inputValue?: FieldPathValue<TFieldValues, TFieldName>
+) {
   batch(() => {
     // Update value state
-    field.setValue(() => value);
+    field.setValue((prevValue) =>
+      field.transform.reduce(
+        (current, transformation) => transformation(current, event),
+        inputValue ?? prevValue
+      )
+    );
 
     // Update touched state
     field.setTouched(true);
@@ -39,9 +50,7 @@ export function updateFieldValue<
     // Update dirty state
     updateFieldDirty(form, field);
 
-    // Validate value if necessary
-    validateIfRequired(form, field, name, {
-      on: ['touched', 'input'],
-    });
+    // Validate value if required
+    validateIfRequired(form, field, name, { on: validationModes });
   });
 }